refactor(lattice-client-react): split lattice config context into separate file
