+++ typescript/packages/lattice-client-react/src/index.ts
@@ -2,7 +2,7 @@
 export * from '@wasmcloud/lattice-client-core';
 
 // package exports
-export {LatticeConfigProvider,LatticeConfigContext} from '@/context/lattice-config-provider';
+export {LatticeConfigProvider} from '@/context/lattice-config-provider';
 export {useLatticeConfig} from '@/use-lattice-config';
 export {useLatticeData} from '@/use-lattice-data';
 export {LatticeClientProvider} from '@/context/lattice-client-provider';
