+++ typescript/packages/lattice-client-react/src/use-lattice-config.ts
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {
   LatticeConfigContext,
   LatticeConfigContextType,
-} from './context/lattice-config-provider';
+} from './context/lattice-config-context';
 
 /**
  * get the current lattice config and a function to update it
